+++ UI/ChallengeDetails/Tabs/Mechanics/Tabs/Damage/Tabs.lua
@@ -18,16 +18,12 @@ Creates tabs (with click-handlers) for Damage frame.
 function MyDungeonsBook:DamageFrame_CreateTabButtonsFrame(parentFrame)
 	local tabs = self:TabsWidget_Create(parentFrame);
 	tabs:SetTabs({
-		{value = "avoidableDamage", text = L["Avoidable Damage"]},
 		{value = "damageDoneToPartyMembers", text = L["All damage taken"]},
 		{value = "damageDoneByPartyMembers", text = L["All damage done"]},
 		{value = "damageDoneToUnits", text = L["Damage Done To Units"]},
 	});
 	tabs:SetCallback("OnGroupSelected", function (container, _, tabId)
 		container:ReleaseChildren();
-		if (tabId == "avoidableDamage") then
-			self.avoidableDamageFrame = self:AvoidableDamageFrame_Create(container, self.activeChallengeId);
-		end
 		if (tabId == "damageDoneToPartyMembers") then
 			self.damageDoneToPartyMembersFrame = self:DamageDoneToPartyMembersFrame_Create(container, self.activeChallengeId);
 		end